requirements.txt
asgiref==3.6.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Django==3.2.16
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.0.0
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
pyTelegramBotAPI==4.7.0
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.26.13
zipp==3.11.0
psycopg2